# internal/adapter/input/http/apidocs/openapi.yaml
openapi: "3.1.0"

info:
  title: relaybox API
  version: "2026-03-20"
  description: |
    relaybox — generic relay hub. Receives messages from monitoring apps (Beszel, Dozzle, etc.)
    and forwards them to external outputs via templates.

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - bearerAuth: []

tags:
  - name: messages
    description: Message management
  - name: inputs
    description: Input management
  - name: outputs
    description: Output management
  - name: health
    description: Health check

paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthz
      tags: [health]
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]

  /inputs/{inputId}/messages:
    post:
      summary: Receive message
      operationId: postMessage
      tags: [messages]
      parameters:
        - $ref: "#/components/parameters/inputId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Input-specific JSON payload
              additionalProperties: true
      responses:
        "201":
          description: Message created
          headers:
            Location:
              description: URL of the created message
              schema:
                type: string
                example: /inputs/beszel/messages/01J...
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageCreatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: List messages for an input
      operationId: listMessages
      tags: [messages]
      parameters:
        - $ref: "#/components/parameters/inputId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /inputs/{inputId}/messages/{messageId}:
    get:
      summary: Get message by ID
      operationId: getMessage
      tags: [messages]
      parameters:
        - $ref: "#/components/parameters/inputId"
        - $ref: "#/components/parameters/messageId"
      responses:
        "200":
          description: Message found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update message status
      operationId: patchMessage
      tags: [messages]
      description: |
        Allowed transition: FAILED → PENDING (manual re-queue).
        All other transitions return 422.
      parameters:
        - $ref: "#/components/parameters/inputId"
        - $ref: "#/components/parameters/messageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING]
              required: [status]
      responses:
        "200":
          description: Message updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  /inputs:
    get:
      summary: List inputs
      operationId: listInputs
      security: []
      tags: [inputs]
      responses:
        "200":
          description: Input list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Input"

  /inputs/{inputId}:
    get:
      summary: Get input by ID
      operationId: getInput
      security: []
      tags: [inputs]
      parameters:
        - $ref: "#/components/parameters/inputId"
      responses:
        "200":
          description: Input found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Input"
        "404":
          $ref: "#/components/responses/NotFound"

  /outputs:
    get:
      summary: List outputs
      operationId: listOutputs
      security: []
      tags: [outputs]
      responses:
        "200":
          description: Output list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Output"

  /outputs/{outputId}:
    get:
      summary: Get output by ID
      operationId: getOutput
      security: []
      tags: [outputs]
      parameters:
        - $ref: "#/components/parameters/outputId"
      responses:
        "200":
          description: Output found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Output"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Input-specific bearer token from config

  parameters:
    inputId:
      name: inputId
      in: path
      required: true
      schema:
        type: string
        example: beszel
    messageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
        example: 01J...
    outputId:
      name: outputId
      in: path
      required: true
      schema:
        type: string
        example: ops-webhook

  schemas:
    MessageStatus:
      type: string
      enum: [PENDING, DELIVERED, FAILED]
      description: |
        Message delivery status.
        Transitions: PENDING → DELIVERED (success), PENDING → FAILED (retries exhausted),
        FAILED → PENDING (manual re-queue via PATCH).

    Message:
      type: object
      properties:
        id:
          type: string
          description: ULID
        inputId:
          type: string
        status:
          $ref: "#/components/schemas/MessageStatus"
        createdAt:
          type: string
          format: date-time
        retryCount:
          type: integer
          description: Number of delivery attempts
        lastAttemptAt:
          type: ["string", "null"]
          format: date-time
      required: [id, inputId, status, createdAt, retryCount]

    MessageCreatedResponse:
      type: object
      properties:
        id:
          type: string
        inputId:
          type: string
        status:
          $ref: "#/components/schemas/MessageStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, inputId, status, createdAt]

    Input:
      type: object
      properties:
        id:
          type: string
      required: [id]

    Output:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [WEBHOOK, SLACK, DISCORD]
        url:
          type: string
        retryCount:
          type: integer
        retryDelayMs:
          type: integer
      required: [id, type, url, retryCount, retryDelayMs]

    Problem:
      type: object
      description: RFC 7807/9457 Problem Details
      properties:
        type:
          type: string
          example: /errors/Unauthorized
        title:
          type: string
          example: Unauthorized
        status:
          type: integer
          example: 401
        detail:
          type: string
          example: "invalid or missing token for input: beszel"
        traceId:
          type: string
      required: [type, title, status, detail]

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: Request body exceeds 1MB limit
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Invalid status transition
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalServerError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"